issue-1046 streaming moved to Controller with test
