+++ strongbox-commons/src/main/resources/application.yaml
@@ -69,7 +69,7 @@ management:
   endpoints:
     web:
       exposure:
-        include: health,info,beans,metrics,trace,scheduledtasks,threaddump,loggers,logfile,logfilestream
+        include: health,info,beans,metrics,trace,scheduledtasks,threaddump,loggers,logfile
       base-path: /api/monitoring
 cacheManagerConfiguration:
   groupConfig:
