+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/logging/LoggingConfiguration.java
@@ -1,0 +1,28 @@
+package org.carlspring.strongbox.controllers.logging;
+
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.boot.actuate.autoconfigure.logging.LogFileWebEndpointProperties;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
+
+/**
+ * @author Przemyslaw Fusik
+ */
+@Configuration
+@EnableConfigurationProperties(LogFileWebEndpointProperties.class)
+public class LoggingConfiguration
+{
+
+    @Bean
+    @ConditionalOnMissingBean
+    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
+    public SseEmitterAwareTailerListenerAdapter tailerListener(SseEmitter sseEmitter)
+    {
+        return new SseEmitterAwareTailerListenerAdapter(sseEmitter);
+    }
+
+}
