+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/logging/SseEmitterAwareTailerListenerAdapter.java
@@ -1,4 +1,4 @@
-package org.carlspring.strongbox.actuator.logfile.stream;
+package org.carlspring.strongbox.controllers.logging;
 
 import java.io.IOException;
 import java.util.Objects;
