+++ strongbox-web-core/src/test/java/org/carlspring/strongbox/config/IntegrationTest.java
@@ -2,6 +2,7 @@ package org.carlspring.strongbox.config;
 
 import org.carlspring.strongbox.MockedRemoteRepositoriesHeartbeatConfig;
 import org.carlspring.strongbox.app.StrongboxSpringBootApplication;
+import org.carlspring.strongbox.controllers.logging.SseEmitterAwareTailerListenerAdapter;
 import org.carlspring.strongbox.cron.services.CronJobSchedulerService;
 import org.carlspring.strongbox.data.CacheManagerTestExecutionListener;
 import org.carlspring.strongbox.rest.client.MockMvcRequestSpecificationProxyTarget;
@@ -15,6 +16,12 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.restassured.config.ObjectMapperConfig;
+import io.restassured.module.mockmvc.config.RestAssuredMockMvcConfig;
+import io.restassured.module.mockmvc.internal.MockMvcFactory;
+import io.restassured.module.mockmvc.internal.MockMvcRequestSpecificationImpl;
+import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.maven.index.updater.ResourceFetcher;
 import org.junit.jupiter.api.parallel.Execution;
@@ -24,11 +31,7 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
+import org.springframework.context.annotation.*;
 import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestExecutionListeners;
@@ -36,14 +39,7 @@ import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import io.restassured.config.ObjectMapperConfig;
-import io.restassured.module.mockmvc.config.RestAssuredMockMvcConfig;
-import io.restassured.module.mockmvc.internal.MockMvcFactory;
-import io.restassured.module.mockmvc.internal.MockMvcRequestSpecificationImpl;
-import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 
 /**
@@ -51,6 +47,7 @@ import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
  * require web server and remote HTTP protocol.
  *
  * @author Alex Oreshkevich
+ * @author Przemyslaw Fusik
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
@@ -62,7 +59,7 @@ import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
 @WithUserDetails("admin")
 @ActiveProfiles("test")
 @TestExecutionListeners(listeners = CacheManagerTestExecutionListener.class,
-                        mergeMode = TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS)
+        mergeMode = TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS)
 @Execution(ExecutionMode.SAME_THREAD)
 public @interface IntegrationTest
 {
@@ -92,7 +89,8 @@ public @interface IntegrationTest
             final ResourceFetcherFactory resourceFetcherFactory = Mockito.mock(ResourceFetcherFactory.class);
 
             Mockito.when(resourceFetcherFactory.createIndexResourceFetcher(ArgumentMatchers.anyString(),
-                                                                           ArgumentMatchers.any(CloseableHttpClient.class)))
+                                                                           ArgumentMatchers.any(
+                                                                                   CloseableHttpClient.class)))
                    .thenReturn(resourceFetcher);
 
             return resourceFetcherFactory;
@@ -103,12 +101,14 @@ public @interface IntegrationTest
         {
             return new MavenMetadataServiceHelper();
         }
-        
+
         @Bean
         @Scope(proxyMode = ScopedProxyMode.TARGET_CLASS, value = ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-        public MockMvcRequestSpecification mockMvc(ApplicationContext applicationContext) {
-            
-            DefaultMockMvcBuilder builder = MockMvcBuilders.webAppContextSetup((WebApplicationContext) applicationContext);
+        public MockMvcRequestSpecification mockMvc(ApplicationContext applicationContext)
+        {
+
+            DefaultMockMvcBuilder builder = MockMvcBuilders.webAppContextSetup(
+                    (WebApplicationContext) applicationContext);
 
             ObjectMapper objectMapper = applicationContext.getBean(ObjectMapper.class);
             ObjectMapperConfig objectMapperFactory = new ObjectMapperConfig().jackson2ObjectMapperFactory((aClass,
@@ -117,13 +117,41 @@ public @interface IntegrationTest
 
             return new MockMvcRequestSpecificationProxyTarget(
                     new MockMvcRequestSpecificationImpl(new MockMvcFactory(builder), config, null, null,
-                            null, null, null, null));
+                                                        null, null, null, null));
         }
 
         @Bean
-        public RestAssuredArtifactClient artifactClient(MockMvcRequestSpecification mockMvc) 
+        public RestAssuredArtifactClient artifactClient(MockMvcRequestSpecification mockMvc)
         {
             return new RestAssuredArtifactClient(mockMvc);
+        }
+
+        @Primary
+        @Bean
+        @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
+        public SseEmitterAwareTailerListenerAdapter testTailerListener(SseEmitter sseEmitter)
+        {
+            return new TestSseEmitterAwareTailerListenerAdapter(sseEmitter);
+        }
+
+        static class TestSseEmitterAwareTailerListenerAdapter
+                extends SseEmitterAwareTailerListenerAdapter
+        {
+
+            public TestSseEmitterAwareTailerListenerAdapter(final SseEmitter sseEmitter)
+            {
+                super(sseEmitter);
+            }
+
+            @Override
+            public void handle(final String line)
+            {
+                super.handle(line);
+                if (sseEmitter != null)
+                {
+                    sseEmitter.complete();
+                }
+            }
         }
 
     }
